Internal Functions

StochasticIntegrals.ForwardCovariance

StochasticIntegrals.ItoIntegral

StochasticIntegrals.ItoSet

StochasticIntegrals.Mersenne

StochasticIntegrals.NumberGenerator

StochasticIntegrals.SimpleCovariance

StochasticIntegrals.SobolGen

StochasticIntegrals.StochasticIntegralsCovariance

Sobol.next!

StochasticIntegrals.brownians_in_use

StochasticIntegrals.correlation

StochasticIntegrals.covariance

StochasticIntegrals.generate_conditioned_distribution

StochasticIntegrals.get_confidence_hypercube

StochasticIntegrals.get_draws

StochasticIntegrals.get_zero_draws

StochasticIntegrals.log_likelihood

StochasticIntegrals.make_covariance_matrix

StochasticIntegrals.pdf

StochasticIntegrals.to_array

StochasticIntegrals.to_dataframe

StochasticIntegrals.to_draws

StochasticIntegrals.update!

StochasticIntegrals.variance

StochasticIntegrals.volatility

Main Structs

StochasticIntegrals.ItoIntegral — Type

ItoIntegral A struct detailing an ito integral. It contains a UnivariateFunction detailing the integrand as well as a symbol detailing an id of the integral's processes.

Usual (and most general) contructor is: ItoIntegral(brownianid::Symbol, f::UnivariateFunction) Convenience constructor for ItoIntegrals where the integrand is a flat function is: ItoIntegral(brownianid::Symbol, variance::Real)

StochasticIntegrals.ItoSet — Type

ItoSet

Creates an ItoSet. This contains :

A correlation matrix of brownian motions.

A vector giving the axis labels for this correlation matrix.

A dict of ItoInterals. Here the keys should be ids for the ito integrals and the values should be ItoIntegrals.

Determine which Brownian processes are used in an array of ItoIntegrals.

StochasticIntegrals.StochasticIntegralsCovariance — Type

StochasticIntegralsCovariance

This is an abstract type which represents structs that represent covariances over spans of time. The concrete instances of this type should support extracting correlations, covariances, volatilities and random number generation.

StochasticIntegrals.ForwardCovariance — Type

ForwardCovariance

Creates an ForwardCovariance struct. This contains :

An Itoset

Time From

Time To

And in the constructor the following items are generated and stored in the object:

A covariance matrix

Labels for the covariance matrix.

The cholesky decomposition of the covariance matrix.

The inverse of the covariance matrix.

The determinant of the covariance matrix.

The constructors are: ForwardCovariance(itoset::ItoSet, from::Real, to::Real; calculatechol::Bool = true, calculateinverse::Bool = true, calculatedeterminant::Bool = true) ForwardCovariance(itoset::ItoSet, from::Union{Date,DateTime}, to::Union{Date,DateTime}) ForwardCovariance(oldForwardCovariance::ForwardCovariance, from::Real, to::Real) ForwardCovariance(old_ForwardCovariance::ForwardCovariance, from::Union{Date,DateTime}, to::Union{Date,DateTime})

StochasticIntegrals.SimpleCovariance — Type

SimpleCovariance

Creates an SimpleCovariance struct. This is a simplified version of ForwardCovariance and is designed for Ito Integrals that are constant (so correlations do not need to be recalculated). This computational saving is the only advantage - ForwardCovariance is more general. It contains the same elements as a ForwardCovariance:

An Itoset

Time From

Time To

And in the constructor the following items are generated and stored in the object:

A covariance matrix

Labels for the covariance matrix.

The cholesky decomposition of the covariance matrix.

The inverse of the covariance matrix.

The determinant of the covariance matrix.

The constructors are: SimpleCovariance(itoset::ItoSet, from::Real, to::Real; calculatechol::Bool = true, calculateinverse::Bool = true, calculate_determinant::Bool = true)

Random Number Generation

StochasticIntegrals.NumberGenerator — Type

A NumberGenerator must always return a vector of uniforms when called with the function next!(::NumberGenerator).
It can contain whatever it wants as long as it does that method.

StochasticIntegrals.Mersenne — Type

A NumberGenerator wrapper for MersenneTwister. This makes pseudorandom numbers.

StochasticIntegrals.SobolGen — Type

A NumberGenerator wrapper for SobolSeq. This makes quasirandom numbers

Sobol.next! — Function

next!(number_generator::NumberGenerator)

This extracts a random draw given a number generator struct (like a SobolGen or a Mersenne).

StochasticIntegrals.get_draws — Function

get_draws(covar::ForwardCovariance; uniform_draw::Array{T,1} = rand(length(covar.covariance_labels_))) where T<:Real
get_draws(covar::ForwardCovariance, num::Integer; twister::MersenneTwister = MersenneTwister(1234), antithetic_variates = false)

get pseudorandom draws from a ForwardCovariance struct. Other schemes (like quasirandom) can be done by inserting quasirandom numbers in as the uniformdraw. If the antitheticvariates control is set to true then every second set of draws will be antithetic to the previous.

StochasticIntegrals.get_zero_draws — Function

get_zero_draws(covar::ForwardCovariance)

get a draw of zero for all ito_integrals. May be handy for bug hunting.

get_zero_draws(covar::ForwardCovariance, num::Integer)

get an array of zero draws for all ito_integrals. May be handy for bug hunting.

Ito Processes

Missing docstring.

Missing docstring for ItoProcess. Check Documenter's build log for details.

Missing docstring.

Missing docstring for evolve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for evolve_covar_and_ito_processes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for make_ito_process_syncronous_time_series. Check Documenter's build log for details.

Missing docstring.

Missing docstring for make_ito_process_non_syncronous_time_series. Check Documenter's build log for details.

Helper Functions

StochasticIntegrals.volatility — Function

volatility(ito::ItoIntegral, on::Union{Date,DateTime})

Get the volatility of an ItoIntegral on a certain date.

volatility(ito::ItoSet, index::Integer, on::Union{Date,DateTime})
volatility(ito::ItoSet, ito_integral_id::Symbol, on::Union{Date,DateTime})

Get volatility of an ito_integral on a date.

volatility(covar::ForwardCovariance, index::Integer, on::Union{Date,DateTime})
volatility(covar::ForwardCovariance, id::Symbol, on::Union{Date,DateTime})

Get the volatility of an ItoIntegral on a date..

StochasticIntegrals.variance — Function

variance(ito::ItoIntegral, from::Real, to::Real)
variance(ito::ItoIntegral, base::Union{Date,DateTime}, from::Union{Date,DateTime}, to::Union{Date,DateTime})

Get the variance of an ItoIntegral from one point of time to another.

variance(covar::ForwardCovariance, id::Symbol)
variance(covar::ForwardCovariance, index::Integer)

Get the variance of an ItoIntegral over a period.

StochasticIntegrals.covariance — Function

covariance(ito1::ItoIntegral,ito2::ItoIntegral, from::Real, to::Real, gaussian_correlation::Real)
covariance(ito1::ItoIntegral,ito2::ItoIntegral, base::Union{Date,DateTime}, from::Union{Date,DateTime}, to::Union{Date,DateTime}, gaussian_correlation::Real)

Get the covariance of two ItoIntegrals over a certain period given the underlying Brownian processes have a correlation of gaussian_correlation.

covariance(covar::ForwardCovariance, index_1::Integer, index_2::Integer)
covariance(covar::ForwardCovariance, id1::Symbol, id2::Symbol)

Get the covariance of two ItoIntegrals over a period.

StochasticIntegrals.correlation — Function

correlation(ito::ItoSet, index1::Integer, index2::Integer)
correlation(ito::ItoSet, brownian_id1::Symbol, brownian_id2::Symbol)

Get correlation between brownian motions in an ItoSet.

correlation(covar::ForwardCovariance, index_1::Integer, index_2::Integer)
correlation(covar::ForwardCovariance, id1::Symbol, id2::Symbol)

Get the correlation of two ItoIntegrals over a period.

StochasticIntegrals.pdf — Function

pdf(covar::ForwardCovariance, coordinates::Dict{Symbol,Real})

get the value of the pdf at some coordinates.

StochasticIntegrals.make_covariance_matrix — Function

make_covariance_matrix(ito_set_::ItoSet, from::Real, to::Real)

Make a covariance matrix given an ItoSet and a period of time. This returns a Hermitian covariance matrix as well as a vector of symbols representing the axis labelling on this Hermitian.

Inputs

ito_set_ - An ItoSet you want to make a covariance matrix from.

from - The (numeric) time from which the covariance span starts.

to - The (numeric) time at which the covariance span ends.

Returns

A Hermitian covariance matrix.

A Vector{Symbol} of labels for the covariance matrix.

StochasticIntegrals.log_likelihood — Function

log_likelihood(covar::ForwardCovariance, coordinates::Dict{Symbol,Real})

get the log likelihood at some coordinates.

StochasticIntegrals.brownians_in_use — Function

brownians_in_use(itos::Array{ItoIntegral,1}, brownians::Array{Symbol,1})

Determine which Browninan processes are used in an array of ItoIntegrals.

StochasticIntegrals.update! — Function

update!(sc::SimpleCovariance, from::Real, to::Real)

This takes a SimpleCovariance and updates it for a new span in time. The new span in time is between from and to. For SimpleCovariance this is done by just adjusting the covariances for the new time span (with corresponding adjustments) to the cholesky, inverse, etc.

The corresponding technique for a ForwardCovariance (which is also a StochasticIntegralsCovariance) is to feed it into a new ForwardCovariance constructor which will recalculate for the new span.

Internal Functions

StochasticIntegrals.to_draws — Function

to_draws(X::Array{T,2}; labels::Array{Symbol,1} = Symbol.("x", 1:size(X)[2]))
to_draws(dd::DataFrame; labels::Array{Symbol,1} =  Symbol.(names(dd)))

Convert array or dataframe to a vector of dicts containing draws.

StochasticIntegrals.to_dataframe — Function

to_dataframe(X::Array{T,2}; labels::Array{Symbol,1} = Symbol.("x", 1:size(X)[2]))
to_dataframe(draws::Array{Dict{Symbol,T},1}; labels::Array{Symbol,1}  = collect(keys(draws[1])))

Convert arrys or vectors of dicts to a dataframe.

StochasticIntegrals.to_array — Function

to_array(draws::Array{Dict{Symbol,T},1}; labels::Array{Symbol,1}  = collect(keys(draws[1]))) where T<:Real
to_array(dd::DataFrame; labels::Array{Symbol,1} =  Symbol.(names(dd)))

Convert draws or a dataframe to an array and a vector of column labels.

StochasticIntegrals.get_confidence_hypercube — Function

get_confidence_hypercube(covar::ForwardCovariance, confidence_level::Real, data::Array{T,2}; tuning_parameter::Real = 1.0)

This returns the endpoints of a hypercube that contains confidence_level% of the dataset.

StochasticIntegrals.generate_conditioned_distribution — Function

generate_conditioned_dist(covar::ForwardCovariance, conditioning_draws::Array{Symbol,T}) where T<:Real

Given some subset of known stochastic integral values this generates a conditional multivariate normal distribution.